requirements.txt
flask
bcrypt
cryptography
mysql-connector-python